refactor duplicated service package
